+++ data/src/main/java/team/retum/data/remote/response/recruitment/FetchRecruitmentsResponse.kt
@@ -9,12 +9,12 @@ data class RecruitmentsResponse(
 )
 
 data class Recruitment(
-    @SerializedName("recruit_id") val recruitId: Int,
+    @SerializedName("recruit_id") val recruitId: Long,
     @SerializedName("company_name") val companyName: String,
     @SerializedName("company_profile_url") val companyProfileUrl: String,
-    @SerializedName("train_pay") val trainPay: Int,
+    @SerializedName("train_pay") val trainPay: Long,
     @SerializedName("military") val military: Boolean,
-    @SerializedName("total_hiring") val totalHiring: Int,
+    @SerializedName("total_hiring") val totalHiring: Long,
     @SerializedName("job_code_list") val jobCodeList: String,
     @SerializedName("bookmarked") val bookmarked: Boolean,
 )
